backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
redis
google-genai
httpx
scenedetect[opencv]
python-dotenv
numpy
opencv-python-headless
aiofiles
ffmpeg-python
